Exceptions: Removes Unnecessary space added by CosmosException.ToString() (#5446)
## Description

The method `ToString()` of `CosmosException` inserts an unusual,
unnecessary space between the type and the message.

#### Expected behavior
The format should be {0}: {1}, without a leading space before the
semicolon

#### Actual behavior
The format is {0} : {1}, with a space

This PR removes the unnecessary space being added to the message.

## Type of change

- [] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #5433

Co-authored-by: Kiran Kumar Kolli <kirankk@microsoft.com>