[Thin Client Integration]: Adds empty string for RegionalDatabaseAccount parameter while creating RNTBD request. (#5471)
# Pull Request Template

## Description

Starting direct version 3.41.0, we will need to pass empty string for
regionalDatabaseAccountName to serialize the request for proxy.
ThinProxy has support of retrieving regionalDatabaseAccountName when its
empty. Instead of sending globalDatabaseAccountName send empty string.

The new contract in TransportSerialization:
https://msdata.visualstudio.com/CosmosDB/_git/CosmosDB/commit/3a7ef7e72685727de7762dd6769042cd765f7397?path=/Product/Microsoft.Azure.Documents/SharedFiles/Rntbd2/TransportSerialization.cs&version=GBmaster&line=112&lineEnd=113&lineStartColumn=1&lineEndColumn=1&type=2&lineStyle=plain&_a=files

## Type of change

Please delete options that are not relevant.

- [X] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #5426