Barrier Requests: Adds Integration tests for retry logic for 410/ Lease Not Found (1022) exceptions (#5469)
# Pull Request Template

## Description

If a barrier request receives a 410/1022 from the backend for any given
replica (quorum replicas in case of reads), then the SDK should attempt
to retry the barrier request in the primary replica. If the primary
replica responds with a 410/1022, then the entire Read or Write
operation should bail-out/ fail and Reads/ Writes (either PPAF enabled
on SM or in MM) should be retried on the next preferred region.

This PR adds integration tests for testing end to end flow with direct
package changes.

## Type of change

Please delete options that are not relevant.

- [X] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #5383