HttpTimeoutPolicy Improvements Phase 3: Fixes the TimeoutPolicies to Have Segregated Timeouts for Point-Reads and Non-Point-Reads with Uniform Timeouts (#5497)
# Pull Request Template

## Description

At a high-level, this PR aims to unify the timeouts across PPAF and thin
client http retry policies for read only requests. Furthermore,
HttpTimeoutPolicies for PPAF and Thinclient will have segregated
timeouts for Point-Reads and Non-Point-Reads. With this change, we will
have the ability to fine-tune/optimize the timeouts for the respective
operations. In this context, requests are allowed to retry 3 times with
6s, 6s and 10s timeouts respectively for each retry. We chose these
timeouts based on the default backend response timeouts and it will help
requests being timeout with aggressive timeouts.

As next step, we have planned for DR drill to identify the appropriate
timeouts. Depending on the findings from the drill, we may go towards
the direction of configurable timeout for first retry.

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #5496