[Internal] Binary Encoding: Adds support to DateTimeOffset type in CosmosDBNewtonsoftWriter (#5474)
## Description

During some of the internal customer (IC3) validation, it was identified
that when the Cosmos .Net SDK is initialized with binary encoding, and a
`DateTimeOffset` field is passed in the payload, the serialization layer
is unable to parse the field and hence fails the operation.

This PR fixes the serialization gap by extending support to
`DateTimeOffset` type in `CosmosDBNewtonsoftWriter`

## Type of change

Please delete options that are not relevant.

- ️✅ Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #5473

---------

Co-authored-by: Debdatta Kunda <87335885+kundadebdatta@users.noreply.github.com>
Co-authored-by: Kiran Kumar Kolli <kirankk@microsoft.com>