[Internal] Pipelines: Adds test ignore category (#5519)
# Pull Request Template

## Description

This PR adds support for the TestCategory `Ignore`. This is needed
temporarily until pipelines are setup to support the preview
SemanticReranking service, so the current pipelines do not fail. The
SemanticReranking preview API is now validated locally and tests are
marked with Ignore category until this happens.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)
## Closing issues

To automatically close an issue: closes #IssueNumber