[Internal] Testing: Refactors `AvailabilityStrategyNoTriggerTest` values to make test more reliable  (#5524)
Changed values to make test more reliable

# Pull Request Template

## Description

This pull request makes minor adjustments to the test parameters in the
`CosmosAvailabilityStrategyTests.cs` file to fine-tune the timing and
thresholds used in the `AvailabilityStrategyNoTriggerTest`. These
changes help to better calibrate the test scenarios for response delays
and hedging thresholds.

Timing and threshold adjustments:

* Reduced the response delay from 300ms to 200ms in one fault injection
scenario to make the test more responsive.
* Increased the response delay from 3000ms to 5000ms in another fault
injection scenario for broader delay coverage.
* Lowered the cross-region hedging threshold from 300ms to 150ms to
trigger hedging sooner in the test.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #IssueNumber