[Internal] Tests: Adds clean-up method to ensure created databases are deleted (#5538)
# Pull Request Template

## Description


This pull request ensures that created databases to the live account are
deleted. On our multimaster test account there are many redundant
databases that have not been cleaned up that can be traced back to this
test. This change ensures that the databases created for this test are
deleted after the test is complete. This should help save on costs.

Resource management improvements:

* Added `client` and `database` as internal instance variables to the
test class, allowing for consistent access and disposal throughout the
test lifecycle.
* Implemented a `[TestCleanup]` method that deletes the test database
(if it exists) and disposes of the Cosmos client, handling exceptions
gracefully to avoid test failures due to cleanup issues.

Test code consistency:

* Updated test methods to assign the created database to `this.database`
and to use the instance variable instead of local variables, ensuring
that the cleanup method can properly dispose of resources.
[[1]](diffhunk://#diff-5b8e740df9df73b4ff1179be9cf42b611977cfa5de6f5f7824abb82fb9a7770fL801-R826)
[[2]](diffhunk://#diff-5b8e740df9df73b4ff1179be9cf42b611977cfa5de6f5f7824abb82fb9a7770fL922-R947)

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #IssueNumber