[Internal] Logging: Refactors CFP traces to contain more information  (#5542)
# Pull Request Template

## Description

mirror of: #5527

If there are errors here that don't directly translate within the
exception, there is useful context that is potentially missing for
debugging here. Making these changes to add some more information that
can be useful in debugging CFP scenarios for users.


## Type of change

Please delete options that are not relevant.

- [] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #IssueNumber